pyproject.toml
[project]
name = "zeno"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.17.2",
    "bcrypt==4.0.1",
    "fastapi[standard]>=0.122.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg>=3.2.13",
    "psycopg-binary>=3.2.13",
    "pydantic-settings>=2.12.0",
    "uuid6>=2025.0.1",
]
